tests/ambient-light/w3c/AmbientLightSensor_browsing_context.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: Sensor readings must only be available in the top-level browsing context</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="support-iframe.html" id="frame" style="display:none" sandbox="allow-scripts">
</iframe>
<script>

async_test(function (t) {
  window.onmessage = t.step_func(function(e) {
    assert_equals(e.data, "SecurityError");
    t.done();
  });
}, "sensor readings can not be fired within iframes");

test(function() {
  let win = window.open('', '_blank');
  try {
    let sensor = new AmbientLightSensor();
    sensor.start();
    assert_not_equals(String(sensor.reading), "[object AmbientLightSensorReading]");
    sensor.stop();
  } catch (e) {
    assert_equals(e.name, "SecurityError");
  } finally {
    win.close();
  }
}, "sensor readings can not be fired on the background tab");

</script>
